+++ R/build-reference-index.R
@@ -14,6 +14,7 @@ data_reference_index <- function(pkg = ".") {
   has_icons <- purrr::some(rows, ~ .x$row_has_icons %||% FALSE)
 
   check_missing_topics(rows, pkg)
+  rows <- Filter(function(x) !x$is_internal, rows)
 
   print_yaml(list(
     pagetitle = tr_("Function reference"),
@@ -23,16 +24,15 @@ data_reference_index <- function(pkg = ".") {
 }
 
 data_reference_index_rows <- function(section, index, pkg) {
-  if (identical(section$title, "internal")) {
-    return(list())
-  }
+  is_internal <- identical(section$title, "internal")
 
   rows <- list()
   if (has_name(section, "title")) {
     rows[[1]] <- list(
       title = section$title,
       slug = make_slug(section$title),
-      desc = markdown_text_block(section$desc, pkg = pkg)
+      desc = markdown_text_block(section$desc, pkg = pkg),
+      is_internal = is_internal
     )
   }
 
@@ -40,7 +40,8 @@ data_reference_index_rows <- function(section, index, pkg) {
     rows[[2]] <- list(
       subtitle = section$subtitle,
       slug = make_slug(section$subtitle),
-      desc = markdown_text_block(section$desc, pkg = pkg)
+      desc = markdown_text_block(section$desc, pkg = pkg),
+      is_internal = is_internal
     )
   }
 
@@ -57,7 +58,8 @@ data_reference_index_rows <- function(section, index, pkg) {
     rows[[3]] <- list(
       topics = purrr::transpose(contents),
       names = names,
-      row_has_icons = !purrr::every(contents$icon, is.null)
+      row_has_icons = !purrr::every(contents$icon, is.null),
+      is_internal = is_internal
     )
   }