+++ tests/testthat/test-build-reference-index.R
@@ -12,14 +12,14 @@ test_that("can generate three types of row", {
 
 test_that("rows with title internal are dropped", {
   ref <- list(
-    list(title = "internal"),
-    list(contents = c("a", "b", "c", "e", "?")),
+    list(title = "internal", contents = c("a", "b")),
+    list(contents = c("c", "e", "?")),
     list(title = "internal")
   )
   meta <- list(reference = ref)
   pkg <- as_pkgdown(test_path("assets/reference"), override = meta)
 
-  index <- data_reference_index(pkg)
+  expect_warning(index <- data_reference_index(pkg), NA)
   expect_equal(length(index$rows), 1)
 })